+++ Contracts/Requests/ItemDiscountRequest.cs
@@ -1,4 +1,4 @@
 namespace BasketManagementAPI.Contracts.Requests;
 
-public sealed record ItemDiscountRequest(string Type, decimal Amount);
+public sealed record ItemDiscountRequest(string Type, int Amount);
