+++ Domain/Shipping/ShippingDetails.cs
@@ -1,4 +1,4 @@
 namespace BasketManagementAPI.Domain.Shipping;
 
-public sealed record ShippingDetails(string Country, decimal Cost);
+public sealed record ShippingDetails(string Country, int Cost);
